splicing: use the generic negotiation mechanism to splice in/out of a channel.
The initiator (in the case of simultaneity, the higher feepayer) pays
for the input and output, and sets the feerate.

We freeze the channel while this construction is going on: it should
be quick.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>


Header from folded patch 'fixup3.patch':

fixup! splicing: use the generic negotiation mechanism to splice in/out of a channel.

- Remove 1 per minute restriction: 25% growth limits us a lot (@niftynei)
- Nomenclature: initiator, splice transaction, channel funding output.
- Weaken reserve check to only cover iff they extract funds.