Use warning instead of disconnecting
When fee ranges are incompatible, disconnecting doesn't help much: upon
reconnection, the same `closing_signed` will be sent again.

It makes more sense instead to send a `warning` and keep the connection open.
The node operator should notice that warning and instruct its node to send
another `closing_signed` with a different fee range (if it agrees) or may
choose to ignore it and wait, or to force-close.