Trim closing output below network dust threshold
When `dust_limit_satoshis` are below 546 sats, we may create closing txs that
won't propagate (e.g. if one of the closing scripts is not using segwit).

This isn't a critical issue since the channel can be force-closed and the
commit should always propagate, but we can do better.

Signers should remove their output if it causes the transaction to fail
propagation. This is backwards-compatible because they already had the
option to remove their own output (and receivers should check the signature
against both potential closing transactions).