accounting: round to nearest whole satoshi for balances, not down
Over the lifetime of a node, rounding every msat balance down (adding to
fees) is a net loss of funds from the network to the miners.

To make things a bit more even, instead we can round to the nearest
satoshi for every peer balance output (on close txs and the
to_local/to_remote balances).

It gets a bit more complicated with HTLC output balances as there is no
'inherent yin/yang' relationship btw the htlc outputs; in other words
you may round up multiple times instead of one rounding up, the other
rounding down, which ends up with bad math.

Consider the following.

You have a 5000msats starting balance.

You add 2 htlcs of 750msats.

Here's what the balances would be if maintained in msats:

5000 - 1500 = 3500 to_local
               750 htlc_1
               750 htlc_2

With the applied rounding rule (round up if over 499), here's what
the following outputs would be:

4sat to_local
1sat htlc_1
1sat htlc_2

4 + 1 + 1 = 6, which is greater than the 5 sats we started with.

So to make this ok, the HTLCs are *always* rounded down, which means
that if the unilateral close they're in ever goes to chain,
there's a good chance you're down a fraction of a satoshi but those
won't be the only satoshis you're out if you go to chain with HTLCs
because recovering those funds onchain has a fee cost of generally
greater than a satoshi etc.