Peer backup storage (features 40/41/42/43)
Nodes can offer to altruistically store small amounts of data on behalf
of their channel peers. It complements `option_data_loss_protect` and can
let nodes that lost data fully recover their channels.

There are two types of backups available:

* Per-channel backups, for channel-related data
* Per-node backups, for data potentially unrelated to channels

There are two distinct pairs of features bits, because this is an
assymmetric feature: one feature bit for the consumers and one for the
providers of peer backups.

This gives a lot of flexibility for use-cases where nodes may want to be
both a consumer and provider of the feature, and lets consumers require
their peers to be providers of the feature.