BOLT#02+07: splicing and changing to simple-taproot-channels
This commit outlines a way to combine the splicing flow with a way
to upgrade the channel-type to a simple-taproot-channel. The new
funding output will be a musig2 funding output. Most of the diff is
adding musig2 nonces to certain messages to ensure that signing for
this new channel will work.