bolt07: advertise upfront fee policies in channel_update
Add the ability for nodes to advertise custom fee policies. A default
value of 1% (for nodes that advertise that they understand the feature
bit) is used to save the network the resources consumed relaying
defaults. The TLV extension is odd to ensure that old nodes can still
parse the extended gossip.