Bolt 1: Clarify the definition of tu fields
Truncated integer fields, e.g. tu16 cannot have leading zeros
judging by the implementations of the parsers, which are used
across the different ln implementations.
The `can` imposed that it is not necessary to truncate the fields
to values of information.

Signed-off-by: Peter Neuroth <pet.v.ne@gmail.com>

Bolt 1: Add where tuXX fields may be used.

As truncated integers are not limited to TLV records with single
values (e.g. onion-routing payload type 8 - payment data) we want
to further clarify where the tuXX fields can be used.

Signed-off-by: Peter Neuroth <pet.v.ne@gmail.com>