bolt2: relax channel_reserve_satoshis requirements
In some cases it can make sense to extend more trust to your
channel counterparty and therefore not require them to always have
something to lose in a breach scenario. This change relaxes the
requirement to a recommendation to accommodate these scenarios.

The main situation where we see demand for this is in the
LSP<->Mobile use case where the LSP does not desire to impose a
channel reserve requirement on the Mobile side. To accomplish this
today, those use cases would explicitly choose to violate the spec.

This change would make this (very legitimate but risky) choice
one that is ultimately still spec-compliant.