Add `option_zero_reserve`
Lightning channels require nodes to lock some of their channel funds
into a channel reserve, which cannot be used for payments made on
that channel. This guarantees that both nodes always have an output
in the commitment transaction, which they will lose if they publish
a revoked commitment.

While this requirement is generally useful, it creates inefficiencies
since that liquidity can't be used to relay payments, and provides a
bad user experience. In some settings, we may want to remove that
channel reserve and allow nodes to use all of their channel funds.

The `option_zero_reserve` feature lets nodes advertise that they
allow their peers to disable the channel reserve on their side,
and may lift the same requirement in response.