Advertise liquidity ads rates
Add the ability to advertise rates at which nodes wish to sell their
liquidity. Buyers can then connect to sellers and request liquidity,
at one of the advertised rates.

This can be used when creating a dual-funded channel, where both
participants contribute to the funding transaction, and one of them
is paid for (some of) their contribution. This can also be used to
splice additional liquidity into existing channels, when splicing is
supported.

We don't add any constraints to the commitment transaction. The lease
duration isn't protected by script opcodes, which means that the seller
could immediately close the channel, or splice funds out. It is up to
the buyer to then blacklist that seller.

We can in the future modify the commitment transactions to restrict the
seller: a dedicated feature bit will let us add this seamlessly if it is
considered useful.