Extensible liquidity ads rates
Add the ability to advertise rates at which nodes wish to sell their
liquidity. Buyers can then connect to sellers and request liquidity,
at one of the advertised rates.

This can be used when creating a dual-funded channel, where both
participants contribute to the funding transaction, and one of them
is paid for (some of) their contribution. This can also be used to
splice additional liquidity into existing channels, when splicing is
supported.

We support multiple types of liquidity leases, and use an extensible
format to allow adding new types of leases in the future. This can be
useful to allow different ways of paying the lease fees, and different
constraints on the seller or buyer side.

The existing lease types don't add any constraints to the commitment
transactions. The lease duration isn't protected by script opcodes,
which means that the seller could immediately close the channel, or
splice funds out. It is up to the buyer to then blacklist that seller.

We can in the future modify the commitment transactions to restrict the
seller: that can be done by introducing a new `lease_type` if it is
considered useful.