fixup! BOLT 12: offers, sixth draft
Make offer_description optional if there's no amount; this is the common "tip" case.

Note that this implies a minimal offer is simply a bech32 encoding of the offer_node_id field, or "lno1" + bech32(0x2242<node_id>).  This might become normalized as the "default offer" to which you can throw sats to a node directly?