Introduce a dedicated `payment_type` field
The previous commits assumed that paying the liquidity fees in a
different way than during the `interactive-tx` session would be
introduced by creating a new `lease_type`. However, it turns out
that in most cases, how the fees are paid is not correlated to the
`lease_type`. We thus introduce a dedicated type for payment options.

We only define one `payment_type`, where the fees are paid from the
buyer's channel balance during the `interactive-tx` session, but it
is easy to introduce different `payment_type`s that reuse the existing
lease type but pay the fees using pending HTLCs for example.