BOLT 3: Fix example keys in test vectors.
All the test vectors use static keys now, which are listed above
already as the local_payment_basepoint and remote_payment_basepoint.
The keys listed here are the pre-static rotated ones: if you use
these, the vectors don't work!

We actually use the same basepoint for the HTLCs, but never spelled it
out.  So do that now, and these are the local/reote htlc keys.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>