Trampoline payment to blinded paths
When paying a Bolt 12 invoice, the payer may use a trampoline node to
relay that payment. The payer simply includes some of the blinded paths
in the payload for the last trampoline node, who will relay to those
blinded paths. That last trampoline node doesn't learn anything about
the final recipient.

Note that this commit depends on Bolt 12: it references a few types that
are introduced in various commits related to Bolt 12 (`blinded_path` and
`blinded_payinfo`), which can be confusing since Bolt 12 spec is still
in progress. I will clean this up once Bolt 12 is finalized.