Extensible liquidity ads rates
Add the ability to advertise rates at which nodes wish to sell their
liquidity. Buyers can then connect to sellers and request liquidity,
at one of the advertised rates.

This can be used when creating a dual-funded channel, where both
participants contribute to the funding transaction, and one of them
is paid for (some of) their contribution. This can also be used to
splice additional liquidity into existing channels, when splicing is
supported.

We introduce an extensible `payment_type` field to control how the
funding fees are paid. The only `payment_type` currently defined pays
funding fees from the buyer's channel balance during the interactive-tx
session, but it is easy to introduce different `payment_type`s (to pay
the fees using HTLCs for example).

We don't add any constraints to the commitment transactions. Sellers
should keep the channel open for at least a month, but there is no way
to guarantee that. The seller could immediately close the channel, or
splice funds out. It is up to the buyer to then blacklist that seller.

Credits to @niftynei for the original design.