Add `channel_creation_fee_sat` to funding rate
Creating a new channel has an additional cost compared to adding
liquidity to an existing channel: the channel will be closed in the
future, which will require paying on-chain fees. Node operators can
include some `channel_creation_fee_sat` to their liquidity ads to
cover some of that future cost.