Trampoline payment to Bolt 12 invoices
When paying a Bolt 12 invoice, the payer may use a trampoline node to
relay that payment. If the recipient doesn't support trampoline, the
payer simply includes some of the blinded paths in the onion payload
for the trampoline node, who will relay to those blinded paths. The
trampoline node doesn't learn anything about the final recipient, it
only learns the blinded paths they use, which should protect their
privacy (unless using empty blinded paths).

If the recipient supports trampoline and the `invoice_request` contains
the trampoline feature bit, the recipient may set it in its invoice. In
that case, the sender can use every node in the blinded path as a
trampoline node and include the encrypted blinded path data in the
trampoline onion. This also provides the opportunity for the sender
to include additional TLVs for the recipient.