Bless the LND SCB force-close request message format
Sending a `channel_reestablish` with `next_commitment_number` of
zero is used in practice to request a peer fail a channel and
broadcast the latest state (for implementations which continue to
ignore the `error` message).

Because its used in practice we should document it to avoid
accidentally writing up incompatible things in the future.