Cap `shared_anchor` amount to 240 satoshis
We cap the amount of the `shared_anchor` to its dust limit of 240 sats
to ensure that only miners can collect the excess from trimmed outputs.

We cannot keep the `shared_anchor` amount at 0 sats all the time: when
the commitment transaction includes some mining fees (because outputs
are trimmed) it is forbidden by `bitcoind`'s default policy to have a
0-value output, otherwise we those outputs would pollute the utxo set
because nobody would ever spend them if the transaction confirms on its
own.